requirements.txt
aiohttp==3.8.5
lxml==4.9.3
pandas==2.0.3
pdf2image==1.17.0
Pillow==9.4.0
Pillow==10.3.0
pyperclip==1.8.2
python_docx==1.1.0
Requests==2.31.0
selenium==4.20.0
